refactor(serializer): #4400 sync_anchor_bits 도 serializer 로 이관 + 헬퍼 가시성 원복
gestell 리뷰 지적. pack_common_attr_bits 만 옮기고 같은 CTRL_HEADER attr 비트
마스크를 직접 들고 있는 sync_anchor_bits 를 document_core 에 남긴 탓에,
vert_rel_to_to_bits/horz_rel_to_to_bits 를 pub(crate) 로 넓혀야 했다.
그 넓힘 자체가 잔여물이므로 함께 옮기고 두 헬퍼를 private 으로 되돌린다.

common_obj_attr_writer.rs 의 serializer::control import 는 이제
pack_common_attr_bits 하나뿐이다.
